backend/pyproject.toml
[project]
name = "lazymeal-backend"
version = "1.0.0"
description = "LazyMeal backend rebuilt with FastAPI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "asyncpg>=0.30.0",
  "httpx>=0.27.0",
  "PyJWT>=2.9.0",
  "pydantic-settings>=2.5.0",
  "redis>=5.0.8"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.pytest.ini_options]
asyncio_mode = "auto"
